requirements.txt
decorator==4.3.0
Django==1.9.6
djangorestframework==3.4.6
networkx==2.1
numpy==1.15.0
scikit-fuzzy==0.3.1
scipy==1.1.0
six==1.11.0
